edit: readme add: dataset download script add: docker container refactor: refactor directory
clean
