+++ .vscode/launch.json
@@ -1,0 +1,49 @@
+{
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "name": "Python: Train DBSherlock",
+            "type": "python",
+            "request": "launch",
+            "program": "${workspaceFolder}/src/main.py",
+            "console": "integratedTerminal",
+            "justMyCode": true,
+            "args": [
+                "--anormly_ratio=2",
+                "--num_epochs=10",
+                "--batch_size=256",
+                "--mode=train",
+                "--dataset=DBS",
+                "--data_path=dataset/processed_dataset",
+                "--input_c=200",
+                "--output_c=200",
+                "--win_size=25",
+                "--step_size=25",
+            ]
+        },
+        {
+            "name": "Python: Test DBSherlock",
+            "type": "python",
+            "request": "launch",
+            "program": "${workspaceFolder}/src/main.py",
+            "console": "integratedTerminal",
+            "justMyCode": true,
+            "args": [
+                "--anormly_ratio=2",
+                "--num_epochs=10",
+                "--batch_size=256",
+                "--mode=test",
+                "--dataset=DBS",
+                "--data_path=dataset/processed_dataset",
+                "--input_c=200",
+                "--output_c=200",
+                "--win_size=25",
+                "--step_size=25",
+                "--pretrained_model=20",
+            ]
+        }
+    ]
+}
