--- utils/utils.py
@@ -1,17 +1,0 @@
-import os
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-from torch.autograd import Variable
-import numpy as np
-
-
-def to_var(x, volatile=False):
-    if torch.cuda.is_available():
-        x = x.cuda()
-    return Variable(x, volatile=volatile)
-
-
-def mkdir(directory):
-    if not os.path.exists(directory):
-        os.makedirs(directory)
